Update the methodology so that the character under the cursor is not the character being read.
This is done by using a lot of abstraction into the one function StringReader.readWhileFunction(), and then forming calls to that function.

This allows a more centralised certainty of behaviour.

Additionall care had to be taken in parse to make sure that this didn't break anything.

The literal argument parser now ignores ending spaces, this is done in the outer implementation. This means that it could be use for matching direct literals - maybe less needed with readWhileRegexp.